Updated Plane Controller/Added Pickups
Finalized the fuel system and synced to the fuel being used in the Pickup Handler. This now allows for fuel pick-ups to adjust the plane's current fuel level.

In addition to that, the new pickup system allows for the creation of new pickup types, as well as said pickup types' values being adjusted on the fly.